Delete confirm

+++ GoalTracker/ViewModels/GoalsViewModel.cs
@@ -26,9 +26,13 @@ namespace GoalTracker.ViewModels
 
         private async void DeleteGoalAsync(object obj)
         {
-            int goalId = (obj as Category).Id;
-            await App.Database.DeleteCategoryAsync(goalId);
-            CurrentGoals = await App.Database.GetCategoriesAsync();
+            bool answer = await Application.Current.MainPage.DisplayAlert("Are you sure you want to delete this goal?", "All Corresponding entries will also be deleted", "Delete", "Cancel");
+            if (answer)
+            {
+                int goalId = (obj as Category).Id;
+                await App.Database.DeleteCategoryAsync(goalId);
+                CurrentGoals = await App.Database.GetCategoriesAsync();
+            }
         }
 
         private async void UpdateGoal(object obj)
